update lab2.4 & lab 3.3

+++ auto-grader/src/main/java/student/testSuite/lab3/problem3/RectangleTester.java
@@ -27,10 +27,10 @@ public class RectangleTester extends BaseTester {
 	 * argument
 	 */
 	
-	public ParameterTesting[] createArgs(long width, long length) {
+	public ParameterTesting[] createArgs(double width, double length) {
 		return new ParameterTesting[] {
-				new ParameterTesting(long.class, FieldName.WIDTH, width)
-				, new ParameterTesting(long.class, FieldName.LENGTH, length)
+				new ParameterTesting(double.class, FieldName.WIDTH, width)
+				, new ParameterTesting(double.class, FieldName.LENGTH, length)
 				};
 	}
 	
@@ -48,8 +48,8 @@ public class RectangleTester extends BaseTester {
 
 	public ITestCase checkFields(int points) {
 		return super.checkFields(points
-				, new FieldTesting(long.class, FieldName.WIDTH)
-				, new FieldTesting(long.class, FieldName.LENGTH)
+				, new FieldTesting(double.class, FieldName.WIDTH)
+				, new FieldTesting(double.class, FieldName.LENGTH)
 				);
 	}
 
@@ -58,7 +58,7 @@ public class RectangleTester extends BaseTester {
 	 */
 
 	public ITestCase constructor(int points) throws ClassNotFoundException {
-		return super.checkConstructorDeclaration(points, long.class, long.class);
+		return super.checkConstructorDeclaration(points, double.class, double.class);
 	}
 	
 	/*
@@ -69,7 +69,7 @@ public class RectangleTester extends BaseTester {
 		return super.methodTester.checkExistence(points, className, method());
 	}
 
-	public ITestCase operateArea(int points, long width, long length, double expected)
+	public ITestCase operateArea(int points, double width, double length, double expected)
 			throws ClassNotFoundException
 			, InstantiationException
 			, IllegalAccessException
